+++ wetterdienst/util/parameter.py
@@ -6,6 +6,12 @@ class _GetAttrMeta(type):
     def __getitem__(cls, x):
         return getattr(cls, x)
 
+    def __iter__(cls):
+        """ Getting subclasses which usually represent resolutions """
+        for attr in vars(cls):
+            if not attr.startswith("_"):
+                yield attr
+
 
 class DatasetTreeCore(metaclass=_GetAttrMeta):
     pass